Broke out tests to separate modules; moved code to aggregation
