+++ pandas/tests/series/apply/test_series_apply.py
@@ -8,7 +8,6 @@ import pandas as pd
 from pandas import DataFrame, Index, MultiIndex, Series, isna
 import pandas._testing as tm
 from pandas.core.base import SpecificationError
-from pandas.core.groupby.base import transformation_kernels
 
 
 class TestSeriesApply:
@@ -210,23 +209,17 @@ class TestSeriesAggregate:
             f_abs = np.abs(string_series)
 
             # ufunc
-            result = string_series.transform(np.sqrt)
+            result = string_series.apply(np.sqrt)
             expected = f_sqrt.copy()
             tm.assert_series_equal(result, expected)
 
-            result = string_series.apply(np.sqrt)
-            tm.assert_series_equal(result, expected)
-
             # list-like
-            result = string_series.transform([np.sqrt])
+            result = string_series.apply([np.sqrt])
             expected = f_sqrt.to_frame().copy()
             expected.columns = ["sqrt"]
             tm.assert_frame_equal(result, expected)
 
-            result = string_series.apply([np.sqrt])
-            tm.assert_frame_equal(result, expected)
-
-            result = string_series.transform(["sqrt"])
+            result = string_series.apply(["sqrt"])
             tm.assert_frame_equal(result, expected)
 
             # multiple items in list
@@ -237,10 +230,6 @@ class TestSeriesAggregate:
             result = string_series.apply([np.sqrt, np.abs])
             tm.assert_frame_equal(result, expected)
 
-            result = string_series.transform(["sqrt", "abs"])
-            expected.columns = ["sqrt", "abs"]
-            tm.assert_frame_equal(result, expected)
-
             # dict, provide renaming
             expected = pd.concat([f_sqrt, f_abs], axis=1)
             expected.columns = ["foo", "bar"]
@@ -249,46 +238,12 @@ class TestSeriesAggregate:
             result = string_series.apply({"foo": np.sqrt, "bar": np.abs})
             tm.assert_series_equal(result.reindex_like(expected), expected)
 
-            expected = pd.concat([f_sqrt, f_abs], axis=1)
-            expected.columns = ["foo", "bar"]
-            result = string_series.transform({"foo": np.sqrt, "bar": np.abs})
-            tm.assert_frame_equal(result, expected)
-
-            # UDF via apply
-            def func(x):
-                if isinstance(x, Series):
-                    raise ValueError
-                return x + 1
-
-            result = string_series.transform(func)
-            expected = string_series + 1
-            tm.assert_series_equal(result, expected)
-
-            # UDF that maps Series -> Series
-            def func(x):
-                if not isinstance(x, Series):
-                    raise ValueError
-                return x + 1
-
-            result = string_series.transform(func)
-            expected = string_series + 1
-            tm.assert_series_equal(result, expected)
-
     def test_transform_and_agg_error(self, string_series):
         # we are trying to transform with an aggregator
-        msg = "Function did not transform"
-        with pytest.raises(ValueError, match=msg):
-            string_series.transform(["min", "max"])
-
         msg = "cannot combine transform and aggregation"
         with pytest.raises(ValueError, match=msg):
             with np.errstate(all="ignore"):
                 string_series.agg(["sqrt", "max"])
-
-        msg = "Function did not transform"
-        with pytest.raises(ValueError, match=msg):
-            with np.errstate(all="ignore"):
-                string_series.transform(["sqrt", "max"])
 
         msg = "cannot perform both aggregation and transformation"
         with pytest.raises(ValueError, match=msg):
@@ -489,76 +444,6 @@ class TestSeriesAggregate:
         with pytest.raises(expected):
             # e.g. Series('a b'.split()).cumprod() will raise
             series.agg(func)
-
-    def test_transform_none_to_type(self):
-        # GH34377
-        df = pd.DataFrame({"a": [None]})
-
-        msg = "Transform function failed.*"
-        with pytest.raises(ValueError, match=msg):
-            df.transform({"a": int})
-
-
-def test_transform_reducer_raises(all_reductions):
-    op = all_reductions
-    s = pd.Series([1, 2, 3])
-    msg = "Function did not transform"
-    with pytest.raises(ValueError, match=msg):
-        s.transform(op)
-    with pytest.raises(ValueError, match=msg):
-        s.transform([op])
-    with pytest.raises(ValueError, match=msg):
-        s.transform({"A": op})
-    with pytest.raises(ValueError, match=msg):
-        s.transform({"A": [op]})
-
-
-# mypy doesn't allow adding lists of different types
-# https://github.com/python/mypy/issues/5492
-@pytest.mark.parametrize("op", [*transformation_kernels, lambda x: x + 1])
-def test_transform_bad_dtype(op):
-    s = pd.Series(3 * [object])  # Series that will fail on most transforms
-    if op in ("backfill", "shift", "pad", "bfill", "ffill"):
-        pytest.xfail("Transform function works on any datatype")
-    msg = "Transform function failed"
-    with pytest.raises(ValueError, match=msg):
-        s.transform(op)
-    with pytest.raises(ValueError, match=msg):
-        s.transform([op])
-    with pytest.raises(ValueError, match=msg):
-        s.transform({"A": op})
-    with pytest.raises(ValueError, match=msg):
-        s.transform({"A": [op]})
-
-
-@pytest.mark.parametrize("use_apply", [True, False])
-def test_transform_passes_args(use_apply):
-    # transform uses UDF either via apply or passing the entire Series
-    expected_args = [1, 2]
-    expected_kwargs = {"c": 3}
-
-    def f(x, a, b, c):
-        # transform is using apply iff x is not a Series
-        if use_apply == isinstance(x, Series):
-            # Force transform to fallback
-            raise ValueError
-        assert [a, b] == expected_args
-        assert c == expected_kwargs["c"]
-        return x
-
-    pd.Series([1]).transform(f, 0, *expected_args, **expected_kwargs)
-
-
-def test_transform_axis_1_raises():
-    msg = "No axis named 1 for object type Series"
-    with pytest.raises(ValueError, match=msg):
-        pd.Series([1]).transform("sum", axis=1)
-
-
-def test_transform_nested_renamer():
-    match = "nested renamer is not supported"
-    with pytest.raises(SpecificationError, match=match):
-        pd.Series([1]).transform({"A": {"B": ["sum"]}})
 
 
 class TestSeriesMap:
